refactor: rename properties classes and update package structure
