+++ src/main/java/com/sprint/mission/discodeit/config/properties/ApiProperties.java
@@ -1,11 +1,11 @@
-package com.sprint.mission.discodeit.config;
+package com.sprint.mission.discodeit.config.properties;
 
 import jakarta.validation.constraints.NotNull;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.validation.annotation.Validated;
 
-@Validated
 @ConfigurationProperties(prefix = "discodeit.api")
+@Validated
 public record ApiProperties(
     @NotNull
     String serverUrl
